feat: Add description view
