+++ natto/Scene/Description/DescriptionViewController.swift
@@ -1,0 +1,18 @@
+//
+//  DescriptionViewController.swift
+//  natto
+//
+//  Created by 佐川晴海 on 2019/04/13.
+//  Copyright © 2019 佐川　晴海. All rights reserved.
+//
+
+import UIKit
+
+class DescriptionViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+}
